Issue #33: Press-and-hold to set starting tach/hobbs to last known ending tach/hobbs.
We do a hack here.  Instead of trying to get an up-to-the-minute value from the service (requires connectivity and latency), or capturing this in the aircraft itself (gets stale quickly), we simply  maintain a hash table of most-recently seen high-water marks for tach and hobbs, and then we update those values from flights as we download them into recents.  So we only have them available if you've viewed recent flights, but it's quick and robust.