Fix Paul's subtle bug where viewing an existing flight while in flight can muck up takeoff time
Indeed, I have slightly crossed wires (good catch, by the way – this was a subtle, subtle bug!!!)

It happens if you are (a) in flight, (b) autodetecting, and (c) you go to view the *details* of an existing flight.  It doesn’t happen if you switch to another app, it doesn’t happen if you view your currency or totals, it doesn’t happen if you view the list of recent flights.

Gory details:

The bug was that I use the same code to display a recent flight as I use to display the new-flight-in-progress.  And when that view is about to appear, I get a notification from iOS that it’s about to appear.  When it’s about to appear, I update the details of the “in the cockpit” section with the current GPS position (although, of course, that particular section is hidden for existing flights).

As part of updating that section, I have a piece of sanity check code: if I see an engine end time but your state is “in the air”, I force the state to “on the ground.”  After all, if you’re still flying after engine shutdown, things might get bad quickly (though it’s more likely that you’re on the highway home after your flight).

So when you view that recent flight, that flight’s details appear, and in doing so, it does the sanity check, sees that your current state is “in the air”, but it has an engine end time, so it sets the state to “on the ground.”

The recent flight is displayed and all is good.  But a few fractions of a second later, a new GPS sample is received, and since it’s above takeoff speed, the system sends a take-off event along with the sample.  This is correctly sent to the new-flight-in-progress, which registers a takeoff event and updates the time.

Turns out this revealed a second even more subtle bug there as well: when a takeoff is detected, my code should have only updated the takeoff time if a takeoff time was not already specified, which should have been a failsafe in the case of the above bug.  I.e., it should only recorded the first takeoff it received.

But the check was instead checking to see if the entire airborne time (i.e., both a first takeoff AND a last landing) was specified.  Normally this works fine: you get a takeoff, then a landing, then when a 2nd takeoff comes, you have both a takeoff and a landing specified and so it knows that it should not update the takeoff time.  And indeed, normally, you would never get two takeoffs in a row without an intervening landing, so this code works flawlessly…until you get two takeoffs in a row.

But the force-to-the-ground code I describe above does NOT create a landing event, and therefore you get exactly what we saw in your telemetry file: multiple takeoffs without intervening landings.

The fix is trivial for both issues: for the first issue, I only update the “in the cockpit” section for new flights, since it’s not even shown for existing flights.  And the fix for the second is to simply look for an known flight start rather than a known flight time.