Issue #350 - Treat all naked dates as if they were UTC
We've long had a hack where we convert the date-of-flight to a pseudo-UTC date that looks correct at the point of submission, but it's a terrible hack for a few reasons, not least of which that the date can shift before that.
To wit: depart SFO on the 24th of Nov at 10pm.  That's 6am on the 25th Zulu.  Arrive LHR at 10am on the 25th.  That date-of-flight now displays as the 25th.
The best solution I can find is to do the same hack, but earlier in the process.  Specifically:
 * Date Picker to ALWAYS work in UTC.  So if you pick the 24th, it's 00:00Z on the 24th, not offset.
 * For incoming dates (i.e., existing flights), the server is not providing the "Z" suffix (which is why they float).  So I tweaked the WSDL parser to treat this as UTC anyhow.
 * We also need to specify UTC when displaying them so that the display date doesn't float around based on your time zone; new dateStringUtc extension method makes this easier.  (Also exposed UTCDateFromLocalDate() extension on Date, mirroring the NSDate extension
 * We were able to remove the old hack from issue #321.  The commit code is now MUCH cleaner - no "stashedDate" to save/restore.  That was just ugly.
 * Similar issue arises with Aircraft maintenance dates, but these already had the "Z" suffix from the server, so it's just necessary to update the date picker.  E.g., "tap for today" if you are in Tokyo at 6am on the 24th was yielding the 23rd.